Refactor: Extract DatasetRunnerStats leaf from DatasetRunner.cpp (P0 split)
Move pure Pearson/Spearman/Kendall/RMSD helpers into DatasetRunnerStats.{h,cpp}
with no I/O or ranking side effects. Document the chunked split plan and wire
CMake for benchmark_datasets + DatasetRunner tests. Full split deferred.
